remove unneded headers
