+++ base/common/LineReader.h
@@ -4,6 +4,9 @@
 
 #include <atomic>
 #include <vector>
+#include <string>
+
+using String = std::string;
 
 class LineReader
 {
