+++ src/Core/MySQL/IMySQLWritePacket.cpp
@@ -41,7 +41,7 @@ size_t getLengthEncodedNumberSize(uint64_t x)
     }
 }
 
-size_t getLengthEncodedStringSize(const String & s)
+size_t getLengthEncodedStringSize(const std::string & s)
 {
     return getLengthEncodedNumberSize(s.size()) + s.size();
 }
@@ -69,13 +69,13 @@ void writeLengthEncodedNumber(uint64_t x, WriteBuffer & buffer)
     }
 }
 
-void writeLengthEncodedString(const String & s, WriteBuffer & buffer)
+void writeLengthEncodedString(const std::string & s, WriteBuffer & buffer)
 {
     writeLengthEncodedNumber(s.size(), buffer);
     buffer.write(s.data(), s.size());
 }
 
-void writeNulTerminatedString(const String & s, WriteBuffer & buffer)
+void writeNulTerminatedString(const std::string & s, WriteBuffer & buffer)
 {
     buffer.write(s.data(), s.size());
     buffer.write(0);
