+++ src/Core/MySQL/PacketEndpoint.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <Core/Types.h>
 #include <IO/ReadBuffer.h>
 #include <IO/WriteBuffer.h>
 #include "IMySQLReadPacket.h"
