+++ src/Interpreters/CrashLog.h
@@ -4,7 +4,7 @@
 
 
 /// Call this function on crash.
-void collectCrashLog(Int32 signal, UInt64 thread_id, const String & query_id, const StackTrace & stack_trace);
+void collectCrashLog(Int32 signal, UInt64 thread_id, const DB::String & query_id, const StackTrace & stack_trace);
 
 
 namespace DB
